lmb-1549: move the form saving logic to the wizard
