+++ app/components/mandataris/edit/form.hbs
@@ -55,104 +55,3 @@
     {{/if}}
   {{/unless}}
 </div>
-
-<AuModal
-  @modalOpen={{this.isSecondModalOpen}}
-  @closable={{true}}
-  @closeModal={{this.cancel}}
->
-  <:title>
-    Bewerk mandaat
-    <p class="au-u-h4 au-u-regular">
-      {{@bestuursorganenTitle}}
-    </p>
-
-  </:title>
-  <:body>
-    <div>
-      <AuLabel @required={{true}}>
-        Reden van aanpassing
-      </AuLabel>
-      <PowerSelect
-        @allowClear={{false}}
-        @selected={{this.reasonForChange}}
-        @options={{this.reasonForChangeOptions}}
-        @onChange={{this.selectReasonForChange}}
-        @searchEnabled={{true}}
-        @noMatchesMessage="Geen resultaten"
-        as |reason|
-      >
-        {{reason}}
-      </PowerSelect>
-    </div>
-  </:body>
-  <:footer>
-    <AuToolbar as |Group|>
-      <Group>
-        <AuButtonGroup>
-          <AuButton
-            @loading={{this.isSaving}}
-            @loadingMessage="Aanpassingen worden bewaard"
-            {{on "click" this.saveForm}}
-          >
-            Opslaan
-          </AuButton>
-          <AuButton {{on "click" this.goBackSecondModal}} @skin="secondary">
-            Ga terug
-          </AuButton>
-        </AuButtonGroup>
-      </Group>
-    </AuToolbar>
-  </:footer>
-</AuModal>
-
-<AuModal
-  @modalOpen={{this.isRangordeModalOpen}}
-  @closable={{true}}
-  @closeModal={{fn (mut this.isRangordeModalOpen) false}}
->
-  <:title>Rangorde vervanger</:title>
-  <:body>
-    <p>
-      U heeft een mandataris, die een rangorde heeft, verhinderd. Deze rangorde
-      kan niet automatisch overgenomen worden in de vervanger. Het toewijzen van
-      een rangorde kan namelijk gevolgen hebben op de rangorde van andere
-      mandatarissen. Wenst u dit nu aan te passen? Het is ook mogelijk dit op
-      een later tijdstip aan te passen in de wijzig rangorde pagina.
-    </p>
-  </:body>
-  <:footer>
-    <AuToolbar as |Group|>
-      <Group>
-        <AuLink
-          @skin="button"
-          @route="organen.orgaan.edit-rangorde"
-          @model={{@bestuursorgaanIT.isTijdsspecialisatieVan.id}}
-          @query={{hash
-            bestuursperiode=@bestuursorgaanIT.heeftBestuursperiode.id
-          }}
-        >
-          Ja
-        </AuLink>
-        <AuButton
-          @skin="secondary"
-          {{on "click" (fn (mut this.isRangordeModalOpen) false)}}
-        >Nee</AuButton>
-      </Group>
-    </AuToolbar>
-  </:footer>
-</AuModal>
-
-<Mandatarissen::Linked::Modal
-  @recentUpdate={{this.selectedReplacement}}
-  @mandataris={{@mandataris.id}}
-  @replacement="true"
-/>
-
-<Mandatarissen::Linked::Modal
-  @recentUpdate={{this.updatedStateMandataris}}
-  @mandataris={{@mandataris.id}}
-  @newMandataris={{this.newMandataris.id}}
-  @callback={{this.callbackAfterUpdate}}
-  @updateState="true"
-/>
