+++ app/components/mandataris/edit/replacement-form.js
@@ -9,7 +9,6 @@ import {
   isDisabledForBestuursorgaan,
   isRequiredForBestuursorgaan,
 } from 'frontend-lmb/utils/is-fractie-selector-required';
-import { getNietBekrachtigdPublicationStatus } from 'frontend-lmb/utils/get-mandataris-status';
 
 export default class MandatarisEditReplacementForm extends Component {
   @service store;
@@ -19,9 +18,6 @@ export default class MandatarisEditReplacementForm extends Component {
   @tracked startDate;
   @tracked endDate;
   @tracked fractie;
-
-  @tracked replacementMandataris;
-  @tracked replacementLidmaatschap;
 
   @tracked overlappingMandate;
   @tracked hasReplacementError;
@@ -78,47 +74,6 @@ export default class MandatarisEditReplacementForm extends Component {
   async updateErrorMap({ id, hasErrors }) {
     this.errorMap.set(id, !!hasErrors);
     this.errorMap = new Map(this.errorMap);
-    const formValues = {};
-    if (!this.formHasErrors) {
-      if (this.replacementMandataris && this.replacementLidmaatschap) {
-        this.replacementMandataris.rollbackAttributes();
-        this.replacementLidmaatschap.rollbackAttributes();
-        this.args.mandataris.tijdelijkeVervangingen = [];
-      }
-
-      const newMandatarisProps = await this.mandatarisService.createNewProps(
-        this.args.mandataris,
-        {
-          start: this.startDate,
-          einde: this.endDate,
-          rangorde: '',
-          publicationStatus: await getNietBekrachtigdPublicationStatus(
-            this.store
-          ),
-        }
-      );
-
-      this.replacementMandataris =
-        await this.mandatarisService.getOrCreateReplacement(
-          this.args.mandataris,
-          this.person,
-          newMandatarisProps,
-          { saveWhenCreated: false }
-        );
-      this.args.mandataris.tijdelijkeVervangingen = [
-        this.replacementMandataris,
-      ];
-      this.replacementLidmaatschap =
-        await this.mandatarisService.createNewLidmaatschap(
-          this.replacementMandataris,
-          this.fractie,
-          { saveWhenCreated: true }
-        );
-      formValues.mandataris = this.args.mandataris;
-      formValues.replacementMandataris = this.replacementMandataris;
-      formValues.replacementLidmaatschap = this.replacementLidmaatschap;
-    }
-
-    this.args.onFormIsValid?.(!this.formHasErrors, formValues);
+    this.args.onFormIsValid?.(!this.formHasErrors, this.person);
   }
 }
